fixing flyway migration error, moving page into resume event and moving page location as per ticket
